requirements.txt
cohere==5.5.4
Flask==3.0.3
Flask-Mail==0.10.0
langchain==0.1.14
langchain-cohere==0.1.5
langchain-community==0.0.31
langchain-core==0.2.4
langchain-text-splitters==0.0.1
marshmallow==3.21.1
mysql-connector-python==8.4.0
pytest==8.1.1
pytest-mock==3.14.0
python-dotenv==1.0.1